# chore(lib): replacing manual `jax.vmap` and broadcasting with vectorize