Using `jnp.vectorize` to reduce the code complexity.

After running the usual benchmarks, as well as a new custom benchmark (see below) to estimate the impact on the compilation time, this patch does not seem to alter the performance at all. However, it greatly reduces the complexity of the code, so I'll merge this.

Timings:

```
hyperfine --warmup 3 'uv run python bench.py' 'uv run python bench.py --skip' 'uv run python bench.py --op=fermat' 'uv run python bench.py --op=fermat --skip'

(this patch)

Benchmark 1: uv run python bench.py
  Time (mean ± σ):      1.293 s ±  0.041 s    [User: 2.136 s, System: 0.331 s]
  Range (min … max):    1.265 s …  1.403 s    10 runs
 
Benchmark 2: uv run python bench.py --skip
  Time (mean ± σ):      1.172 s ±  0.011 s    [User: 2.005 s, System: 0.320 s]
  Range (min … max):    1.157 s …  1.197 s    10 runs
 
Benchmark 3: uv run python bench.py --op=fermat
  Time (mean ± σ):      2.020 s ±  0.016 s    [User: 2.900 s, System: 0.374 s]
  Range (min … max):    2.001 s …  2.045 s    10 runs
 
Benchmark 4: uv run python bench.py --op=fermat --skip
  Time (mean ± σ):      1.364 s ±  0.066 s    [User: 2.203 s, System: 0.338 s]
  Range (min … max):    1.288 s …  1.479 s    10 runs
 
(before changes)

Benchmark 1: uv run python bench.py
  Time (mean ± σ):      1.335 s ±  0.095 s    [User: 2.150 s, System: 0.339 s]
  Range (min … max):    1.276 s …  1.541 s    10 runs
 
Benchmark 2: uv run python bench.py --skip
  Time (mean ± σ):      1.161 s ±  0.006 s    [User: 2.013 s, System: 0.313 s]
  Range (min … max):    1.155 s …  1.174 s    10 runs
 
Benchmark 3: uv run python bench.py --op=fermat
  Time (mean ± σ):      2.001 s ±  0.009 s    [User: 2.899 s, System: 0.370 s]
  Range (min … max):    1.986 s …  2.016 s    10 runs
 
Benchmark 4: uv run python bench.py --op=fermat --skip
  Time (mean ± σ):      1.291 s ±  0.012 s    [User: 2.161 s, System: 0.333 s]
  Range (min … max):    1.275 s …  1.312 s    10 runs
```

`bench.py`:

```python
import argparse

import jax

from differt.rt import fermat_path_on_linear_objects, image_method

if __name__ == "__main__":
    key = jax.random.PRNGKey(1234)
    parser = argparse.ArgumentParser(
        description="Benchmark image_method or fermat_path_on_linear_objects"
    )
    parser.add_argument(
        "--op",
        choices=["image_method", "fermat"],
        default="image_method",
    )
    parser.add_argument("-s", "--skip", action="store_true")

    args = parser.parse_args()

    if args.op == "image_method":
        key_from, key_to, key_mirror, key_mirror_normals = jax.random.split(key, 4)
        from_vertex = jax.random.normal(key_from, (3,))
        to_vertex = jax.random.normal(key_to, (3,))
        mirror_vertices = jax.random.normal(key_mirror, (3, 3))
        mirror_normals = jax.random.normal(key_mirror_normals, (3, 3))
        if not args.skip:
            image = image_method(
                from_vertex, to_vertex, mirror_vertices, mirror_normals
            ).block_until_ready()
    else:
        key_from, key_to, key_object_origins, key_object_vectors = jax.random.split(
            key, 4
        )
        from_vertex = jax.random.normal(key_from, (3,))
        to_vertex = jax.random.normal(key_to, (3,))
        object_origins = jax.random.normal(key_object_origins, (3, 3))
        object_vectors = jax.random.normal(key_object_vectors, (3, 2, 3))

        if not args.skip:
            fermat_path = fermat_path_on_linear_objects(
                from_vertex, to_vertex, object_origins, object_vectors
            ).block_until_ready()
```